fix(frontend): keep export dialog sections refreshable
